+++ src/server/.babelrc.js
@@ -1,13 +1,12 @@
 'use strict';
 
-const plugins = ['@babel/transform-runtime', 'transform-require-extensions'];
-
-if (process.env.BABEL_ESM)
-  plugins.push(require('../../babelPluginAddBabelRuntimeFileExtensions.js'));
-
 module.exports = {
   comments: false,
-  plugins,
+  plugins: [
+    '@babel/transform-runtime',
+    'transform-runtime-file-extensions',
+    'transform-require-extensions',
+  ],
   presets: [
     [
       '@babel/env',